+++ packages/partials/async/src/Functions/WithTimeout.ts
@@ -17,18 +17,40 @@
 import { TimeoutError } from '../Errors';
 
 /**
+ * The extended options for `withTimeout` function.
+ */
+export interface IWithTimeoutOptions<T> {
+
+    /**
+     * The callback to receive the result of the asynchronous task after timeout.
+     *
+     * @param error     The error occurred if the task timed out, otherwise null.
+     * @param result    The result of the task if it completed in time.
+     */
+    collectResult?: (error: unknown | null, result?: T) => void;
+}
+
+/**
  * Bind a timeout to an asynchronous task.
  *
  * > The real asynchronous task is still running even if the timeout is reached.
  * > If you wanna cancel/abort the task, you should consider using `AbortTimeoutController` instead.
+ * >
+ * > Besides, you may need to handle the result of the task even after timeout by yourself,
+ * > just use `IWithTimeoutOptions.collectResult` for that.
  *
- * @param ms            The timeout duration in milliseconds.
+ * @param timeoutMs     The timeout duration in milliseconds.
  * @param asyncTask     The asynchronous task to execute, which can be a Promise or a function that returns a Promise.
+ * @param opts          Additional options for handling the task result.
  *
  * @throws {TimeoutError} If the asynchronous task does not complete within the specified timeout.
  * @returns A Promise with the result of the asynchronous task or a timeout error.
  */
-export async function withTimeout<T>(ms: number, asyncTask: Promise<T> | (() => Promise<T>)): Promise<T> {
+export async function withTimeout<T>(
+    timeoutMs: number,
+    asyncTask: Promise<T> | (() => Promise<T>),
+    opts?: IWithTimeoutOptions<T>,
+): Promise<T> {
 
     const pr: Promise<T> = typeof asyncTask === 'function' ? asyncTask() : asyncTask;
 
@@ -37,7 +59,7 @@ export async function withTimeout<T>(ms: number, asyncTask: Promise<T> | (() => 
         let timer: NodeJS.Timeout | null = setTimeout(() => {
             timer = null; // Clear the timer reference
             reject(new TimeoutError(pr));
-        }, ms);
+        }, timeoutMs);
 
         pr.then(
             (result) => {
@@ -45,14 +67,18 @@ export async function withTimeout<T>(ms: number, asyncTask: Promise<T> | (() => 
                     clearTimeout(timer);
                     timer = null;
                     resolve(result);
+                    return;
                 }
+                opts?.collectResult?.(null, result);
             },
             (error) => {
                 if (timer) {
                     clearTimeout(timer);
                     timer = null;
                     reject(error);
+                    return;
                 }
+                opts?.collectResult?.(error);
             },
         );
     });